+++ Craftsman/Builders/Projects/FunctionalTestsCsProjBuilder.cs
@@ -39,13 +39,13 @@ public class FunctionalTestsCsProjBuilder
     <PackageReference Include=""FluentAssertions"" Version=""6.6.0"" />
     <PackageReference Include=""MediatR"" Version=""10.0.1"" />
     <PackageReference Include=""Moq"" Version=""4.17.2"" />
-    <PackageReference Include=""NUnit"" Version=""3.13.3"" />
-    <PackageReference Include=""NUnit3TestAdapter"" Version=""4.2.1"" />
     <PackageReference Include=""Microsoft.AspNetCore.Mvc.Testing"" Version=""6.0.4"" />
     <PackageReference Include=""Microsoft.AspNetCore.Mvc.NewtonsoftJson"" Version=""6.0.4"" />
     <PackageReference Include=""Microsoft.NET.Test.Sdk"" Version=""17.1.0"" />
     <PackageReference Include=""Respawn"" Version=""5.0.1"" />
     <PackageReference Include=""WebMotions.Fake.Authentication.JwtBearer"" Version=""6.0.0"" />
+    <PackageReference Include=""xunit"" Version=""2.4.1"" />
+    <PackageReference Include=""xunit.runner.visualstudio"" Version=""2.4.5"" />
   </ItemGroup>
 
   <ItemGroup>