+++ Craftsman/Builders/Tests/FunctionalTests/GetEntityListTestBuilder.cs
@@ -42,7 +42,7 @@ public class GetEntityListTestBuilder
 using {fakerClassPath.ClassNamespace};
 using {testUtilClassPath.ClassNamespace};{permissionsUsing}
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 using System.Net;
 using System.Threading.Tasks;
 
@@ -60,7 +60,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
 
         _client.AddAuth(new[] {{Roles.SuperAdmin}});" : "";
 
-        return $@"[Test]
+        return $@"[Fact]
     public async Task {testName}()
     {{
         // Arrange
@@ -77,7 +77,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
     private static string GetEntityTestUnauthorized(Entity entity)
     {
         return $@"
-    [Test]
+    [Fact]
     public async Task get_{entity.Name.ToLower()}_list_returns_unauthorized_without_valid_token()
     {{
         // Arrange
@@ -94,7 +94,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
     private static string GetEntityTestForbidden(Entity entity)
     {
         return $@"
-    [Test]
+    [Fact]
     public async Task get_{entity.Name.ToLower()}_list_returns_forbidden_without_proper_scope()
     {{
         // Arrange